miscellaneous improvements and addressed comments
