+++ Cargo.toml
@@ -17,7 +17,7 @@ color-eyre = "0.6.3"
 crossterm = { version = "0.29.0" }
 ratatui = { version = "0.29.0", features = ["unstable-widget-ref"] }
 smol_str = "0.2.2"
-soft_ratatui = { version = "0.0.2", optional = true }
+soft_ratatui = { version = "0.0.7", optional = true }
 tracing = "0.1.41"
 
 [dev-dependencies]
@@ -33,7 +33,7 @@ opt-level = 1
 opt-level = 3
 
 [features]
-default = ["std", "async_executor", "windowed"]
+default = ["std", "async_executor"]
 windowed = ["dep:soft_ratatui", "bevy/default"]
 
 # Features for `std` platforms
