+++ examples/kitty.rs
@@ -2,6 +2,7 @@ use bevy::{
     app::{AppExit, ScheduleRunnerPlugin},
     prelude::*,
 };
+
 use bevy_ratatui::{
     RatatuiPlugins, event::KeyEvent, kitty::KittyEnabled, terminal::RatatuiContext,
 };
