+++ src/crossterm.rs
@@ -1,0 +1,59 @@
+use std::io::{self, Stdout, stdout};
+
+use bevy::{app::AppExit, prelude::*};
+
+use crossterm::{
+    ExecutableCommand, cursor,
+    terminal::{EnterAlternateScreen, LeaveAlternateScreen, disable_raw_mode, enable_raw_mode},
+};
+use ratatui::Terminal;
+
+use crate::terminal::*;
+use crate::{kitty::KittyEnabled, mouse::MouseCaptureEnabled};
+use ratatui::backend::CrosstermBackend;
+
+/// A wrapper around ratatui::Terminal that automatically enters and leaves the alternate screen.
+///
+/// This resource is used to draw to the terminal. It automatically enters the alternate screen when
+/// it is initialized, and leaves the alternate screen when it is dropped.
+///
+/// # Example
+///
+/// ```rust
+/// use bevy::prelude::*;
+/// use bevy_ratatui::terminal::RatatuiContext;
+///
+/// fn draw_system(mut context: ResMut<RatatuiContext>) {
+///     context.draw(|frame| {
+///         // Draw widgets etc. to the terminal
+///     });
+/// }
+/// ```
+
+impl TerminalContext for RatatuiContext {
+    fn init() -> io::Result<Self> {
+        let mut stdout = stdout();
+        stdout.execute(EnterAlternateScreen)?;
+        enable_raw_mode()?;
+        let backend = CrosstermBackend::new(stdout);
+        let terminal = Terminal::new(backend)?;
+        Ok(RatatuiContext(terminal))
+    }
+
+    fn restore() -> io::Result<()> {
+        let mut stdout = stdout();
+        stdout
+            .execute(LeaveAlternateScreen)?
+            .execute(cursor::Show)?;
+        disable_raw_mode()?;
+        Ok(())
+    }
+}
+
+impl Drop for RatatuiContext {
+    fn drop(&mut self) {
+        if let Err(err) = Self::restore() {
+            eprintln!("Failed to restore terminal: {}", err);
+        }
+    }
+}
